in lasers and radio-waves, switched from Vector2D.Float to Vector2D.Double, see #2467, and removed unused code I came across

+++ trunk/simulations-java/simulations/radio-waves/src/edu/colorado/phet/radiowaves/model/movement/ManualMovement.java
@@ -26,7 +26,7 @@ public class ManualMovement implements MovementType {
     private double[] yAHistory = new double[s_posHistoryLength - 2];
     private float vAve;
     private float aAve;
-    private Vector2D.Float velocity = new Vector2D.Float();
+    private Vector2D.Double velocity = new Vector2D.Double();
     private MedianFilter dataFilter = new MedianFilter( yVHistory );
 
     /**
